Added camera struct

+++ go-rt/internal/maths/size2.go
@@ -2,5 +2,5 @@ package maths
 
 type Size2 struct {
 	Width float64
-	Height float32
+	Height float64
 }
